精進+ストレステスト環境を用意

+++ problems/abc169/abc169_c.py
@@ -1,5 +1,6 @@
+from decimal import Decimal
+
 A, B = input().split()
 A = int(A)
-B = int(float(B)*100)
-print(A*B//100)
-
+B = int(Decimal(B) * 100)
+print((A*B)//100)
